+++ chaincode/main_test.go
@@ -18,7 +18,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"math/rand"
 	"os"
 	"strings"
 	"testing"
@@ -314,11 +313,3 @@ func TestQueryEmptyResponse(t *testing.T) {
 }
 
 var characterRunes = []rune("abcdef0123456789")
-
-func GetRandomHash() string {
-	b := make([]rune, 64)
-	for i := range b {
-		b[i] = characterRunes[rand.Intn(len(characterRunes))]
-	}
-	return string(b)
-}
