+++ chaincode/traintuple_composite.go
@@ -224,9 +224,6 @@ func (traintuple *CompositeTraintuple) Save(db *LedgerDB, traintupleKey string) 
 		if err := db.CreateIndex("computePlan~computeplanid~worker~rank~key", []string{"computePlan", traintuple.ComputePlanID, traintuple.Dataset.Worker, strconv.Itoa(traintuple.Rank), traintupleKey}); err != nil {
 			return err
 		}
-		if err := db.CreateIndex("computeplan~id", []string{"computeplan", traintuple.ComputePlanID}); err != nil {
-			return err
-		}
 	}
 	if traintuple.Tag != "" {
 		err := db.CreateIndex("compositeTraintuple~tag~key", []string{"compositeTraintuple", traintuple.Tag, traintupleKey})
